Correctly export .dat54 files for BANK-enabled builds
Please note, this breaks Release audio files completely!
If you'd like to export for both Release and Bank builds, export to XML and use a different version of CodeWalker for Release audio files.

Note: This does not update the NameTableOffset, which you'll probably want to fix if you care about the names of the files.

PS: The code is extremely hacky at times, and while it could be refactored to be nicer, this works, and I just need it to work for now.